requirements.txt
Flask
Flask-SQLAlchemy
Flask-Marshmallow
PyJWT